MOZ-2554: Merge fixes from mbed-os-6.13.0-jfd-fixes (#14)
* Changed to C++20

* applied patch from fep low power

* Changed to C++20

* Updated compile options

* Added more fixes

* Added event queue changes

* Changed to hard float and fixed sleep

* Added werror

* Corrected compile warnings/errors

* Updated system clock settings

* Reverted change

* Downgraded STM32Cube to 1.23.0

* Corrected build issues

* Corrected mimimal printf compile warning

* Revert "Downgraded STM32Cube to 1.23.0"

This reverts commit ae1907138ed5d1c980ec736fb2b6d4101bfdfbe5.

* Corrected build issues

* Added I2C slave changes

* Updated from old branch

* Corrected I2C slave command

* Minor rename

* Corrected name

* Virtual i2c read write

Made i2c read/write functions virtual so that they can be overridden.

* Removed obsolete file

* Corrected unused variable in mbed_board.c

* Added maybe_unused to validate_errno_values

* Fix build errors

+++ platform/source/minimal-printf/mbed_printf_implementation.c
@@ -298,7 +298,7 @@ static void mbed_minimal_formatted_string_double(char *buffer, size_t length, in
         /* get decimal part */
         double precision = 1.0;
 
-        for (size_t index = 0; index < dec_precision; index++) {
+        for (int index = 0; index < dec_precision; index++) {
             precision *= 10;
         }
